ACTUAL WORKING CODE WITH A MAIN CLASS !!!
