+++ CA3/Main.h
@@ -1,0 +1,12 @@
+#ifndef MAIN_H
+#define MAIN_H
+
+
+
+class Main {
+
+};
+
+
+
+#endif //MAIN_H
